requirements.txt
Flask
requests
google-generativeai
PyPDF2
python-docx
firebase_admin